backend/requirements.txt
Flask
Flask-Cors
paho-mqtt
sqlite3